Moving global functions to serializers/util
