+++ MHacks/v1/serializers/util.py
@@ -1,7 +1,5 @@
 from rest_framework import serializers
 
-from MHacks.v1.util import to_utc_epoch
-
 
 class UnixEpochDateField(serializers.DateTimeField):
     def to_internal_value(self, value):
@@ -43,3 +41,32 @@ class DurationInSecondsField(serializers.Field):
 
     def to_representation(self, value):
         return value.total_seconds()
+
+
+def parse_date_last_updated(request):
+    date_last_updated_raw = request.query_params.get('since', None)
+    if date_last_updated_raw:
+        try:
+            from pytz import utc
+            from datetime import datetime
+
+            return datetime.utcfromtimestamp(float(date_last_updated_raw)).replace(tzinfo=utc)
+        except ValueError:
+            pass
+    return None
+
+
+def now_as_utc_epoch():
+    from django.utils.timezone import now
+
+    return to_utc_epoch(now())
+
+
+def to_utc_epoch(date_time):
+    from datetime import datetime
+
+    if isinstance(date_time, datetime):
+        from calendar import timegm
+
+        return timegm(date_time.timetuple())
+    return None
